+++ CHANGES.md
@@ -7,6 +7,8 @@ This version of pyfakefs does not support Python 2.6.  Python 2.6 users
 must use pyfakefs 3.3 or earlier.
 
 #### New Features
+  * Added possibility to add own fake modules to `Patcher` / `TestCase`
+   via modules_to_patch argument (see [#345](../../issues/345))
   * Added support for file descriptor path parameter in `os.scandir` (Python >= 3.7, Posix only) (see [#346](../../issues/346))
   * Added support to fake out backported `scandir` module ([#332](../../issues/332))
   * Dynamic loading of modules after setup is now on by default and no more considered experimental (see [#340](../../issues/340))